server-python/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
docker==6.1.3
sqlalchemy==2.0.15
pydantic==2.0.3 